+++ src/ffi_gen_tmp.h
@@ -6,19 +6,32 @@
 // and run:
 //
 // flutter pub run ffigen --config ffigen.yaml
-// to generate [flutter_soloud_FFIGEN.dart]
+// to generate [flutter_soloud_FFIGEN.dart], or if it complains for some reason:
+// export CPATH="$(clang -v 2>&1 | grep "Selected GCC installation" | rev | cut -d' ' -f1 | rev)/include";  dart run ffigen --config ffigen.yaml
 //
 // the generated code will be placed into flutter_soloud_FFIGEN.dart
 // copy the generated definition into flutter_soloud_bindings_ffi.dart
 
 #include "enums.h"
+#include <stdbool.h>
 
 #define FFI_PLUGIN_EXPORT
 
 //--------------------- copy here the new functions to generate
 
-FFI_PLUGIN_EXPORT enum PlayerErrors getFilterParams(
-    unsigned int handle,
-    enum FilterType filterType,
-    int attributeId,
-    float *filterValue);
+FFI_PLUGIN_EXPORT enum ReadSamplesErrors readSamplesFromFile(
+    const char *filePath,
+    float startTime,
+    float endTime,
+    unsigned long numSamplesNeeded,
+    bool average,
+    float *pSamples);
+
+FFI_PLUGIN_EXPORT enum ReadSamplesErrors readSamplesFromMem(
+    const unsigned char *buffer,
+    unsigned long dataSize,
+    float startTime,
+    float endTime,
+    unsigned long numSamplesNeeded,
+    bool average,
+    float *pSamples);